+++ package.json
@@ -9,7 +9,8 @@
     "lint": "eslint .",
     "preview": "vite preview",
     "test": "vitest",
-    "coverage": "vitest run --coverage"
+    "coverage": "vitest run --coverage",
+    "typecheck": "tsc -w --noEmit -p ./tsconfig.app.json"
   },
   "dependencies": {
     "react": "^19.0.0",
